+++ src/main/java/domain/attendance/AttendanceType.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.attendance;
 
 public enum AttendanceType {
