+++ src/main/java/domain/attendance/AttendanceTypeCount.java
@@ -1,6 +1,7 @@
-package domain;
+package domain.attendance;
 
-import java.util.Comparator;
+import domain.date.AttendanceDate;
+import domain.date.AttendanceDateTime;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
