+++ src/main/java/domain/attendance/PenaltyType.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.attendance;
 
 import java.util.Arrays;
