+++ src/main/java/domain/crew/Crew.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.crew;
 
 import java.util.Objects;
