+++ src/main/java/service/AttendanceService.java
@@ -5,14 +5,14 @@ import controller.dto.AttendanceHistoryWithPenaltyTypeDto;
 import controller.dto.AttendanceRequestDto;
 import controller.dto.AttendanceTypeCountDto;
 import controller.dto.AttendanceUpdateResultDto;
-import domain.AttendanceDate;
-import domain.AttendanceDateTime;
-import domain.AttendanceHistories;
-import domain.AttendanceHistory;
-import domain.AttendanceTypeCount;
-import domain.Crew;
-import domain.Crews;
-import domain.PenaltyType;
+import domain.date.AttendanceDate;
+import domain.date.AttendanceDateTime;
+import domain.attendance.AttendanceHistories;
+import domain.attendance.AttendanceHistory;
+import domain.attendance.AttendanceTypeCount;
+import domain.crew.Crew;
+import domain.crew.Crews;
+import domain.attendance.PenaltyType;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
