+++ src/main/java/view/OutputView.java
@@ -4,8 +4,8 @@ import controller.dto.AttendanceHistoryDto;
 import controller.dto.AttendanceHistoryWithPenaltyTypeDto;
 import controller.dto.AttendanceTypeCountDto;
 import controller.dto.AttendanceUpdateResultDto;
-import domain.AttendanceType;
-import domain.PenaltyType;
+import domain.attendance.AttendanceType;
+import domain.attendance.PenaltyType;
 import java.util.Comparator;
 import java.util.List;
 import java.util.Map;
