+++ src/test/java/domain/AttendanceDateTest.java
@@ -1,6 +1,7 @@
 package domain;
 
 
+import domain.date.AttendanceDate;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
