+++ src/test/java/domain/CrewsTest.java
@@ -1,9 +1,7 @@
 package domain;
 
-import static org.junit.jupiter.api.Assertions.*;
-
+import domain.crew.Crews;
 import org.assertj.core.api.Assertions;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
