+++ src/test/java/domain/PenaltyTypeTest.java
@@ -1,14 +1,11 @@
 package domain;
 
-import static org.junit.jupiter.api.Assertions.*;
-
+import domain.attendance.PenaltyType;
 import java.util.stream.Stream;
 import org.assertj.core.api.Assertions;
-import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
 import org.junit.jupiter.params.provider.MethodSource;
-import org.junit.jupiter.params.provider.ValueSource;
 
 class PenaltyTypeTest {
